training/pyproject.toml
[project]
name = "final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "kaggle>=1.7.4.5",
    "numpy<2",
    "opencv-python-headless<4.12",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "tqdm>=4.67.1",
    "ultralytics>=8.3.253",
    "wandb>=0.24.0",
    # QAT (Quantization-Aware Training) 의존성
    # NOTE: pytorch-quantization은 NVIDIA NGC 저장소에만 존재합니다.
    # 수동 설치 필요: uv pip install pytorch-quantization --extra-index-url https://pypi.ngc.nvidia.com
    "onnx>=1.14.0",
    "onnxsim>=0.4.33",
]

[project.optional-dependencies]
qat = [
    "pytorch-quantization>=2.1.3",
]

[[tool.uv.index]]
name = "nvidia"
url = "https://pypi.ngc.nvidia.com"
explicit = true
